pyproject.toml
[project]
name = "forge-runtime"
version = "0.1.0"
description = "Integrated container runtime + embedded workflow orchestration. Lightning-fast alternative to Docker + Airflow."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Christopher Torres", email = "torresjchristopher@github.com"}
]
license = {text = "MIT"}
keywords = ["containers", "orchestration", "workflow", "dag", "airflow", "docker-alternative"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Software Distribution",
    "Topic :: Utilities",
]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10",
    "psutil>=5.9",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "pytest-mock>=3.10",
]

[project.scripts]
forge = "forge.cli:main"

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["forge"]

[tool.black]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=forge --cov-report=term-missing"
